+++ source/extensions/omni.isaac.orbit_tasks/omni/isaac/orbit_tasks/utils/wrappers/rsl_rl/vecenv_wrapper.py
@@ -165,7 +165,9 @@ class RslRlVecEnvWrapper(VecEnv):
         obs = obs_dict["policy"]
         extras["observations"] = obs_dict
         # move time out information to the extras dict
-        extras["time_outs"] = truncated
+        # this is only needed for infinite horizon tasks
+        if not self.unwrapped.cfg.is_finite_horizon:
+            extras["time_outs"] = truncated
 
         # return the step information
         return obs, rew, dones, extras