+++ source/extensions/omni.isaac.orbit_tasks/test/wrappers/test_rsl_rl_wrapper.py
@@ -92,6 +92,41 @@ class TestRslRlVecEnvWrapper(unittest.TestCase):
             print(f">>> Closing environment: {task_name}")
             env.close()
 
+    def test_no_time_outs(self):
+        """Check that environments with finite horizon do not send time-out signals."""
+        for task_name in self.registered_tasks[0:5]:
+            print(f">>> Running test for environment: {task_name}")
+            # create a new stage
+            omni.usd.get_context().new_stage()
+            # parse configuration
+            env_cfg: RLTaskEnvCfg = parse_env_cfg(task_name, use_gpu=self.use_gpu, num_envs=self.num_envs)
+            # change to finite horizon
+            env_cfg.is_finite_horizon = True
+
+            # create environment
+            env = gym.make(task_name, cfg=env_cfg)
+            # wrap environment
+            env = RslRlVecEnvWrapper(env)
+
+            # reset environment
+            _, extras = env.reset()
+            # check signal
+            self.assertNotIn("time_outs", extras, msg="Time-out signal found in finite horizon environment.")
+
+            # simulate environment for 10 steps
+            with torch.inference_mode():
+                for _ in range(10):
+                    # sample actions from -1 to 1
+                    actions = 2 * torch.rand(env.action_space.shape, device=env.unwrapped.device) - 1
+                    # apply actions
+                    extras = env.step(actions)[-1]
+                    # check signals
+                    self.assertNotIn("time_outs", extras, msg="Time-out signal found in finite horizon environment.")
+
+            # close the environment
+            print(f">>> Closing environment: {task_name}")
+            env.close()
+
     """
     Helper functions.
     """